Fixes issue that didn't let player change activity.
The logic in engine's run was bad which made it so if an activity
finished for whatever reason, the player would set itself as inactive.
Now, we only set the player as inactive if there are no more activities
to do. Later, I'll add the feature to switch players after a certain
time.